requirements.txt
phidata
openai
yfinance
python-dotenv
lancedb
tantivy
pypdf
sqlalchemy
httpx
duckdb
duckduckgo-search
fastapi
uvicorn